refactored functions in messageHandler, moved getUserByUserId to different file, fixed dependency and added lookup for username in database
