+++ project/backend/controllers/userControllers.js
@@ -5,7 +5,7 @@ import bcrypt from 'bcrypt';
 import crypto from 'crypto';
 import pump from 'pump';
 import fs from 'fs';
-import * as utils from '../../utils/utils.js';
+import * as utils from '../utils/utils.js';
 import path from 'path';
 
 // Returns all users registered in the db (userId and username only)
